Terrain Update
Updated plains graphics. Added graphics for desert & tundra. Added graphics variations (1-3). To-do: randomize which image to use.

Added svg files for terrain, but they are of larger file size than png so not sure if they are going to be used.

All tiles should now also have the class tile_ground in case it is a ground tile of any kind.